+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html),
 with the exception that 0.x versions can break between minor versions.
 
-## Unreleased
+## [0.10.0] - 2022-04-28
 ### Added
 - Support for `\G` ([anchor to end of previous match](https://www.regular-expressions.info/continue.html)): Using a regex
   like `\G\w` will match each letter of `foo` in `foo bar` but
@@ -140,6 +140,8 @@ with the exception that 0.x versions can break between minor versions.
 ### Added
 - Initial release
 
+
+[0.10.0]: https://github.com/fancy-regex/fancy-regex/compare/0.9.0...0.10.0
 [0.9.0]: https://github.com/fancy-regex/fancy-regex/compare/0.8.0...0.9.0
 [0.8.0]: https://github.com/fancy-regex/fancy-regex/compare/0.7.1...0.8.0
 [0.7.1]: https://github.com/fancy-regex/fancy-regex/compare/0.7.0...0.7.1